pyproject.toml
[project]
name = "codesize-proxy-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.4.1",
    "titanoboa>=0.2.7",
]
